+++ plc4go/protocols/cbus/readwrite/model/MediaTransportControlDataSelectionName.go
@@ -114,7 +114,7 @@ func (m *_MediaTransportControlDataSelectionName) GetLengthInBitsConditional(las
 	lengthInBits := uint16(m.GetParentLengthInBits())
 
 	// Simple field (selectionName)
-	lengthInBits += uint16(int32(int32(int32(m.GetCommandTypeContainer().NumBytes())-int32(int32(1)))) * int32(int32(8)))
+	lengthInBits += uint16(int32((int32(m.GetCommandTypeContainer().NumBytes()) - int32(int32(1)))) * int32(int32(8)))
 
 	return lengthInBits
 }
