+++ plc4go/protocols/cbus/readwrite/model/TelephonyDataDivert.go
@@ -114,7 +114,7 @@ func (m *_TelephonyDataDivert) GetLengthInBitsConditional(lastItem bool) uint16 
 	lengthInBits := uint16(m.GetParentLengthInBits())
 
 	// Simple field (number)
-	lengthInBits += uint16(int32(int32(int32(m.GetCommandTypeContainer().NumBytes())-int32(int32(1)))) * int32(int32(8)))
+	lengthInBits += uint16(int32((int32(m.GetCommandTypeContainer().NumBytes()) - int32(int32(1)))) * int32(int32(8)))
 
 	return lengthInBits
 }
